+++ common/collections_service.py
@@ -87,6 +87,7 @@ def save_filter_collection(
     rating: str = "All",
     rating_or_higher=False,
     sort_by: str = "Alpha",
+    order_by: str = "Descending",
 ) -> None:
     manager = get_collections_manager()
     manager.add_filter_collection(
@@ -99,5 +100,6 @@ def save_filter_collection(
         rating,
         "true" if is_truthy(rating_or_higher) else "false",
         sort_by,
+        order_by,
     )
     manager.save()