+++ RELEASE_NOTES.md
@@ -1,10 +1,11 @@
 ## VPinFE Release Notes
 
 ### Summary
-This is the offical release of vpinfe v2.0.8
+This is the offical release of vpinfe v2.0.9
 
 ### What's New
-- update score_parser.py
+- Collections now have a filter for sort by runtime and order by
+- add updating .vpx and directb2s files on the fly in the managerui
 
 ### Fixes
 None